% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrayfuns.r
\name{array2binom}
\alias{array2binom}
\title{Convert array to binomial}
\usage{
array2binom(x)
}
\arguments{
\item{x}{An integer array}
}
\value{
binomial array of the same dimension and size as x
}
\description{
Convert an integer array to binomial. Also works with a data frame or tibble.
}
