% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrayfuns.r
\name{dim_drop0}
\alias{dim_drop0}
\title{Drop slices that sum to zero in specified dimension}
\usage{
dim_drop0(x, dropdim)
}
\arguments{
\item{x}{An array whose elements can be summed}

\item{dropdim}{Dimension for which zero-sum slices should be dropped}
}
\value{
array of the same dimension as x
}
\description{
Drop slices of an array that sum to zero in specified dimension (only test on
2D and 3D arrays)
}
