% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_crcmnid.r
\name{import_crcmnid}
\alias{import_crcmnid}
\title{Import CRC MN ID data}
\usage{
import_crcmnid(datdir, mnidfile)
}
\arguments{
\item{datdir}{Directory path for MN ID database file}

\item{mnidfile}{File path for MN ID database file}
}
\value{
list with sightings, loccode, and biopsy sample tables
}
\description{
Import data from MN ID database, including humpback sightings, location code 
reference table, and biopsy sample table
}
