% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_CAPTURE.r
\name{run_CAPTURE}
\alias{run_CAPTURE}
\title{Mark-recapture estimates from CAPTURE}
\usage{
run_CAPTURE(ch, keep.out = FALSE)
}
\arguments{
\item{ch}{Data frame or 2D array with binomial capture histories}

\item{keep.out}{If TRUE, copy capture.out to working directory}
}
\value{
list of estimates, standard errors, and capture probabilities
}
\description{
Obtain mark-recapture estimates from the CAPTURE program for a given set of 
capture histories
}
\details{
This function runs the CAPTURE program (a compiled executable written in 
FORTRAN) to estimate abundance from the provided set of capture histories, 
which code individual animals (rows) as sighted (1) or not (0) in each occasion 
(cols). The CAPTURE program (Rexstad and Burnham 1992; White et al., 1978) was 
developed as a companion to Otis et al.'s (1978) mark-recapture monograph. 

run_CAPTURE was modified from code written by Jay Barlow (Calambokidis and 
Barlow 2020), and currently only runs the Chao Mth (Chao et al., 1992) and 
Darroch Mt models, but can easily be modified to run other estimators available 
in CAPTURE (see program manual in the inst/capture directory of this package).

Beware that run_CAPTURE has not been made robust to incorrect inputs and is not
parallelizable in its current form.
}
\references{
Calambokidis, J., and J. Barlow. 2020. Updated abundance estimates for blue 
and humpback whales along the U.S. West Coast using data through 2018. U.S. 
Department of Commerce, NOAA Technical Memorandum NMFS-SWFSC-634.

Chao, A., S.-M. Lee, and S.L. Jeng. 1992 Estimating population size for 
capture-recapture data when capture probabilities vary by time and individual 
animal. Biometrics 48: 201-216.

Otis, D. L., K. P. Burnham, G. C. White, and D. R. Anderson. 1978. Statistical 
inference from capture data on closed animal populations. Wildlife Monographs 
62. 135 pp. 

Rexstad, E., and K.P. Burnham. 1992. Users Guide for Interactive Program CAPTURE.
Colorado Cooperative Fish & Wildlife Research Unit, Colorado State University, 
Fort Collins, Colorado. 

White, G.C., K.P. Burnham, D.L. Otis, and D.R. Anderson. 1978. Users Manual for 
Program CAPTURE, Utah State Univ. Press, Logan, Utah.
}
