\name{cmrutils-package}
\alias{cmrutils-package}
\alias{cmrutils}
\docType{package}
\title{
  Misc Functions of the Center for the Mathematical Research
}
\description{
  A collection of useful helper routines developed by students of the
  Center for the Mathematical Research, Stankin, Moskow.
}
\details{
\tabular{ll}{
  Package: \tab cmrutils\cr
  Type: \tab Package\cr
  Version: \tab 1.2\cr
  Date: \tab 2009-05-25\cr
  License: \tab GPL (>= 3)\cr
}
}
\author{
  Andrey Paramonov <cmr.Pent@gmail.com>
}
\keyword{ package }
\keyword{ utilities }
