% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_effort.r
\name{import_effort}
\alias{import_effort}
\title{Import effort data}
\usage{
import_effort(datdir, filename = NULL, effyrs = NULL)
}
\arguments{
\item{datdir}{Directory path for effort data files}

\item{filename}{Years for which to import effort}

\item{effyrs}{Years for which to import effort}
}
\value{
list of data frames for with effort, effort events, and event codes
}
\description{
Import data from CRC or SPLASH Sighting databases
}
\details{
Assumes effort data files (e.g., 'CRC Sighting Data YYYY.mdb') all stored in one directory
}
